pyproject.toml
[project]
name = "ib-bench"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openpyxl>=3.1.5",
    "pyyaml>=6.0.3",
    "anthropic>=0.40.0",
    "openai>=1.50.0",
]
